requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
slack-bolt==1.18.1
httpx==0.25.2
sqlmodel==0.0.14
pydantic==2.5.0
jinja2==3.1.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6
aiohttp==3.9.1
openai==1.3.0
